requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-py==0.42.1
alpaca-trade-api==3.2.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.10.0
astroid==3.3.11
attrs==25.3.0
beautifulsoup4==4.13.5
black==25.9.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.117.1
feedparser==6.0.12
finnhub-python==2.4.25
Flask==3.1.2
frozendict==2.4.6
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.3
multidict==6.6.4
multitasking==0.0.12
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.3.3
openai==1.109.1
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pylint==3.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
scipy==1.16.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sseclient-py==1.8.0
starlette==0.48.0
streamlit==1.50.0
ta==0.11.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.36.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
yarl==1.20.1
yfinance==0.2.58
